Draft-04-conforming "additionalProperties"
We formerly treated "additionalProperties" as literally an
additional "properties" object that would get merged in to the
regular one.  The actual specification is that it is a single
schema that is used with any property that neither appears
in "properties" nor matches a pattern in "patternProperties".

As part of this, we would only merge "additionalProperties"
into the properties within the *Of keywords, so instead
apply that logic to "properties" so all such "additionalProperties"
usage can be switched to "properties" and conform with draft-04.

Also support boolean values for "additionalProperties".

Add a regression test for the old behavior.

The schema1 fixture was using a oneOf that could never pass,
because any instance that matched either of them would match both.
Changed this to make sense.  It doesn't really affect the tests,
but it was bothering me.